pyproject.toml
[tool.poetry]
name = "mobile-at-qaguru-lesson-19"
version = "0.1.0"
description = ""
authors = ["Eugene Evdokimneko <eugene.qa94@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
selene = "2.0.0rc9"
requests = "^2.32.3"
pytest = "^8.3.2"
allure-pytest = "^2.13.5"
appium-python-client = "4.0.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



[tool.pytest.ini_options]
addopts = "--clean-alluredir --alluredir=allure-results"